requirements.txt
airflow-multi-dagrun==2.2.0
apache-airflow==2.2.0
apache-airflow-providers-amazon==2.3.0
apache-airflow-providers-apache-spark==2.0.0
apache-airflow-providers-ftp==2.0.1
apache-airflow-providers-http==2.0.1
apache-airflow-providers-imap==2.0.1
apache-airflow-providers-mysql==2.1.1
apache-airflow-providers-slack==4.1.0
apache-airflow-providers-sqlite==2.0.1
boto3==1.18.58
botocore==1.21.58
docker==5.0.3
Jinja2==2.11.3
pandas==1.3.2
requests==2.26.0
dateutil

